Part_2/pyproject.toml
[tool.poetry]
name = "quotes"
version = "0.1.0"
description = ""
authors = ["hlushchenko"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
psycopg2-binary = "^2.9.5"
django-environ = "^0.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
